Replace JCEProviderInfo.java with a simpler one
